Introduce recursive-call macro.
For the better part of a year and a half, the S16X4 development environment
lacked support for making recursive subroutine calls.  However, STS will
require a new calling convention that supports recursion as well as a more
robust means of passing parameters into STS procedures.  I introduce a frame
pointer concept with this push, whose variable is named %fp and is located at
$FFFE at present, and a new macro, callr,, which relies on it.  This breaks
ground for STS procedures and applications to collaborate without relying on
difficult-to-manage global variables.

The intended calling convention is illustrated below:

 %fp+0	Return PC
 %fp+2  param 1
 %fp+4  param 2
 %fp+6  param 3
 %fp+8  result

Note that this convention also permits returning more than one result.  The
callER allocates the space on its frame stack, and is similarly responsible for
cleaning it up afterwards as well.  The callED procedure is able to return to
the caller via a simple %fp @ @ go sequence.