Return AST for expressions, JAL insn.
To support JAL, I needed to support forward references.  This required
expression() to return the AST for its expression, rather than an
immediately computed value.  However, the ADVANCE mnemonic expects an
immediate value, because it needs to alter the location counter.  So I
added an expression evaluator as well, along with a convenience
function, constantExpression(), which behaves exactly like expression()
except that (1) it returns the immediate value of its results, and (2)
throws an exception if the return type isn't an integer.