Initial RV64S support in emulator.
The code is ugly, but functional.  It doesn't support interrupts yet, but that
shouldn't be TOO hard to implement.  However, before I do, I definitely want to
rethink how to support interrupt dispatching.  Right now, traps are
synchronous.  So, if I do support interrupts in the obvious way, then a race
condition can occur: 1) if an interrupt fires, and 2) that results in the need
to trap, then the trap handler will eat the "cause" of the interrupt, and the
ISR will never actually get a chance to run.

How funny -- this is exactly the same bug that the 6502 had with respect to IRQ
and BRK.  Hah!