Use SPI slave as alternate way to print characters.
This is only to prove that I have correct SPI implementation in the
emulator.  It is not intended to replace the debugger interface.
(Although, I kind of wish that I'd done this from the very beginning.
It would eliminate the need for a fake debugger UART interface.)

Eventually, this interface will be used to interact with a fake SD/MMC
card.