Better packaging.
If linked against an application, the application can find the BIOS jump
table by reading location 8 in physical memory.  Character-oriented
devices are now supported with the following API:

- BIOS_CANOUT: Returns true if specified device can accept another byte.

- BIOS_CHROUT: Send byte to indicated device.

- BIOS_CANINP: Returns true if specified device has data to be received.

- BIOS_CHRINP: Receives a byte from the indicated device, if any.

- BIOS_STROUT: Sends a NULL-terminated string of text to the indicated
  device.

- BIOS_I_STROUT: Sends a NULL-terminated string placed inline to the
  BIOS JAL to the indicated device.

Character device 0 is the screen output device.  It emulates a simple
printer that obeys ASCII control codes.  No VT-anything emulation exists
with this simple device.

Character device 1 is reserved for an ASCII-only input stream from the
keyboard, but I haven't implemented that logic yet.  Once completed,
this would be sufficient to port STS to run under the BIOS.

Block-oriented devices are not yet supported.  This will come after
keyboard support.

The only thing that I regret is that I have to write several different
BIOS implementations.  This version obviously targets the e emulator.  e
will draw its raw scancodes and ASCII support from the SDL library, so
it will have much of its work done for it already.  When it comes time
to port to live hardware, the Digilent Nexys2 uses a PS/2-style keyboard
interface.  Thus, the BIOS will need to be changed to support the PS/2
interface.