Support KIA-II (aka EKIA) in BIOS.
Raw KIA-II support in BIOS provides access to SDL's keycodes.  This will
permit me to finally use the SDL window for keyboard entry, and to not
have to pay attention to the shell window all the time.

(And I *still* don't need interrupt support for the KIA-II yet.  But,
in case you were wondering, the CPU's interrupt input will be tied to
the KIA-II's QE flag.)

Of course, programs should ALWAYS rely on BIOS' ASCII keyboard interface
(which doesn't exist yet; that's next), and not the hardware-specific
raw keyboard interface.  But, it's here in case you need it.