Program to show a framebuffer dump, along with a sample image.
I generated the image from some artwork I'd commissioned for the
project.  Loaded it into Gimp, exported it as a C .h file.  Then, used
"cvtimg" to convert it into a form that is compatible with the MGIA
framebuffer format (just have to switch pixel order and color indices,
basically).  Added it to the ROM filesystem constructed by `fs0` in the
STS source code, and away we go.